Refactored folder structure to fit common practices
