+++ App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AppContainer from "./components/Screens/AppContainer";
+import AppContainer from "./src/navigation/AppContainer";
 
 export default function App() {
   return (
