+++ src/components/BarcodeScannerExample.tsx
